Added blue line images ... Implemented blue line table

+++ CineMetro/GreenDetailsViewController.m
@@ -222,7 +222,7 @@ NSMutableArray *points;
     {
         SLComposeViewController *tweetSheetOBJ = [SLComposeViewController
                                                   composeViewControllerForServiceType:SLServiceTypeTwitter];
-        NSString *twittertext = [NSString stringWithFormat:@"#CineMetro #line2station%li\n",position1+1];
+        NSString *twittertext = [NSString stringWithFormat:@"#CineMetro #line2station%li\n",(long)position1+1];
         [tweetSheetOBJ setInitialText:twittertext];
         [self presentViewController:tweetSheetOBJ animated:YES completion:nil];
     }
@@ -236,7 +236,7 @@ NSMutableArray *points;
 - (void)facebookButton{
     if([SLComposeViewController isAvailableForServiceType:SLServiceTypeFacebook]) {
         SLComposeViewController *fbSheetOBJ = [SLComposeViewController composeViewControllerForServiceType:SLServiceTypeFacebook];
-        NSString *fbtext = [NSString stringWithFormat:@"#CineMetro #line2station%li\n",position1+1];
+        NSString *fbtext = [NSString stringWithFormat:@"#CineMetro #line2station%li\n",(long)position1+1];
         [fbSheetOBJ setInitialText:fbtext];
         [self presentViewController:fbSheetOBJ animated:YES completion:Nil];
     }
