+++ CineMetro/RedDetailsViewController.m
@@ -121,7 +121,7 @@ NSMutableArray *points;
     {
         SLComposeViewController *tweetSheetOBJ = [SLComposeViewController
                                                   composeViewControllerForServiceType:SLServiceTypeTwitter];
-        NSString *twittertext = [NSString stringWithFormat:@"#CineMetro #line1station%li\n",indexPath+1];
+        NSString *twittertext = [NSString stringWithFormat:@"#CineMetro #line1station%li\n",(long)indexPath+1];
         [tweetSheetOBJ setInitialText:twittertext];
         [self presentViewController:tweetSheetOBJ animated:YES completion:nil];
     }
@@ -135,7 +135,7 @@ NSMutableArray *points;
 - (void)facebookButton {
     if([SLComposeViewController isAvailableForServiceType:SLServiceTypeFacebook]) {
         SLComposeViewController *fbSheetOBJ = [SLComposeViewController composeViewControllerForServiceType:SLServiceTypeFacebook];
-        NSString *fbtext = [NSString stringWithFormat:@"#CineMetro #line1station%li\n",indexPath+1];
+        NSString *fbtext = [NSString stringWithFormat:@"#CineMetro #line1station%li\n",(long)indexPath+1];
         [fbSheetOBJ setInitialText:fbtext];
         [self presentViewController:fbSheetOBJ animated:YES completion:Nil];
     }
